tests/wpt/web-platform-tests/xhr/getallresponseheaders.htm

<!doctype html>
<title>XMLHttpRequest: getAllResponseHeaders()</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id="log"></div>
<script>
async_test((t) => {
  const client = new XMLHttpRequest()
  client.onload = t.step_func_done(() => {
    assert_equals(client.getAllResponseHeaders(), "foo-test: 1, 2, 3\r\n")
  })
  client.onerror = t.unreached_func("unexpected error")
  client.open("GET", "resources/headers-basic.asis")
  client.send(null)
})

async_test((t) => {
  const client = new XMLHttpRequest()
  client.onload = t.step_func_done(() => {
    assert_equals(client.getAllResponseHeaders(), "also-here: Mr. PB\r\newok: lego\r\nfoo-test: 1, 2\r\n")
  })
  client.onerror = t.unreached_func("unexpected error")
  client.open("GET", "resources/headers.asis")
  client.send(null)
})

test(() => {
  const client = new XMLHttpRequest
  client.open("GET", "resources/header-content-length.asis", false)
  client.send()
  assert_equals(client.getAllResponseHeaders(), "content-length: 0\r\n")
})

async_test(t => {
  const client = new XMLHttpRequest();
  client.onload = t.step_func_done(() => {
    assert_equals(client.getAllResponseHeaders(), "www-authenticate: 1, 2, 3, 4\r\n");
  });
  client.onerror = t.unreached_func("unexpected error");
  client.open("GET", "resources/headers-www-authenticate.asis");
  client.send();
});
</script>
